+++ shadow/changelog.md
@@ -41,3 +41,4 @@ v 0.0.1
 - 解决右边滚动条
 - 解决dashbaord 太丑的问题
 - 调整 dashboard 布局, 先添加一个配置框
+- 调整header border